requirements.txt
django == 4.0.4
djangorestframework == 3.13.1
django-cors-headers == 3.12.0
pillow == 9.0.1
djangorestframework_simplejwt == 5.2.0
django-taggit == 3.0.0